requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
bokeh==2.4.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
fonttools==4.42.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
jsonschema==4.18.1
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==3.20.3
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
rich==13.4.2
rpds-py==0.8.10
six==1.16.0
smmap==5.0.0
streamlit==1.24.1
streamlit-scrollable-textbox==0.0.3
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
xyzservices==2023.5.0
zipp==3.16.0